add search_in_rotated_sorted_array.cpp

+++ 1_linear_table/remove_duplicates_from_sorted_array.cpp
@@ -1,6 +1,6 @@
 /*
 Given a sorted array, remove the duplicates in place such that each element appear only
-once and return the new length. Do not allocate extra space for another array, you must do this in place with constant memory. 
+once and return the new length. Do not allocate extra space for another array, you must do this in place with constant memory.
 For example, Given input array A = [1,1,2], Your function should return length = 2, and A is now [1,2].
 */
 
@@ -18,7 +18,7 @@ public:
             return nums.size();
 
         int index = 0;
-        for (int i = 1; i < nums.size(); i++)
+        for (size_t i = 1; i < nums.size(); i++)
         {
             if (nums[i] != nums[index])
                 nums[++index] = nums[i];
